refactor: cria arquio para cada classe de personagem
